+++ src/Symfony/Component/EventDispatcher/Attribute/AsEventListener.php
@@ -20,16 +20,18 @@ namespace Symfony\Component\EventDispatcher\Attribute;
 class AsEventListener
 {
     /**
-     * @param string|null $event      The event name to listen to
-     * @param string|null $method     The method to run when the listened event is triggered
-     * @param int         $priority   The priority of this listener if several are declared for the same event
-     * @param string|null $dispatcher The service id of the event dispatcher to listen to
+     * @param string|null       $event      The event name to listen to
+     * @param string|null       $method     The method to run when the listened event is triggered
+     * @param int               $priority   The priority of this listener if several are declared for the same event
+     * @param string|null       $dispatcher The service id of the event dispatcher to listen to
+     * @param array|string|null $events     The event or events name to listen to
      */
     public function __construct(
         public ?string $event = null,
         public ?string $method = null,
         public int $priority = 0,
         public ?string $dispatcher = null,
+        public array|string|null $events = null,
     ) {
     }
 }