+++ src/Symfony/Component/EventDispatcher/Tests/Fixtures/TaggedMultiListener.php
@@ -13,9 +13,11 @@ namespace Symfony\Component\EventDispatcher\Tests\Fixtures;
 
 use Symfony\Component\EventDispatcher\Attribute\AsEventListener;
 
-#[AsEventListener(event: CustomEvent::class, method: 'onCustomEvent')]
-#[AsEventListener(event: 'foo', priority: 42)]
-#[AsEventListener(event: 'bar', method: 'onBarEvent')]
+#[AsEventListener(events: CustomEvent::class, method: 'onCustomEvent')]
+#[AsEventListener(events: 'foo', priority: 42)]
+#[AsEventListener(events: 'bar', method: 'onBarEvent')]
+#[AsEventListener(events: [CustomEvent::class, 'foo', 'bar'], method: 'onMultipleEvents', priority: 10)]
+#[AsEventListener(events: [CustomEvent::class, BizEvent::class])]
 final class TaggedMultiListener
 {
     public function onCustomEvent(CustomEvent $event): void
@@ -30,8 +32,21 @@ final class TaggedMultiListener
     {
     }
 
-    #[AsEventListener(event: 'baz')]
+    #[AsEventListener(events: 'baz')]
     public function onBazEvent(): void
+    {
+    }
+
+    public function onMultipleEvents(): void
+    {
+    }
+
+    #[AsEventListener]
+    public function onMultipleEventsWithSomeInvalids(BizEvent|CustomEvent|string|int $event): void
+    {
+    }
+
+    public function __invoke()
     {
     }
 }